Apply Chrome specific privacy settings when running under nw.js
Chrome has enabled some privacy settings per default, which may affect the security of the NanoWallet when running the standalone client under nw.js. This pull request disables multiple settings which are not needed by the standalone client. The `passwordSavingEnabled` setting which is enabled by the default may also cause security issues because it stores automatically the passwords entered into the wallet.

The following settings will be disabled:
- chrome.privacy.services.alternateErrorPagesEnabled
- chrome.privacy.services.autofillEnabled
- chrome.privacy.services.hotwordSearchEnabled
- chrome.privacy.services.passwordSavingEnabled
- chrome.privacy.services.safeBrowsingExtendedReportingEnabled
- chrome.privacy.services.searchSuggestEnabled
- chrome.privacy.services.spellingServiceEnabled
- chrome.privacy.services.translationServiceEnabled

A complete list of the privacy settings can be found on this page: https://developer.chrome.com/extensions/privacy

The pull request updates also nw.js to 0.25.4 and it orders the dependencies in the package.json file.